fix: Passing ID to "Actions" column
## Issue
An `undefined` value  was being passed in as props to the "Actions" column in the "Update columns definition" section. This is because the `accessor` was _only_ passing in the email value so the `cell` method had no knowledge of the ID of the associated item. Also, the ID value lives on the `item.value` not the `item`.