Add missing top padding for card
When the card has a header, the spacing between the header and the content should not be doubled. But when the card has no header, it should have a top padding.

This is achieved by setting the bottom padding of the header to 0, instead of setting the top padding of the card content to 0.

This is symmetrical to how it is done for the footer.