Update command.svelte
Exposed the reflective api of command components so that users can take advantage of it.

I was implementing a text editor and needed to do something like: `commandRef.ref?.updateSelectedByItem(1);` but I quickly discovered that the `ref` is just a node of the dom and not the api itself. I hope this helps others too.